# backend/openapi.yaml
openapi: 3.0.3
info:
  title: GADAVIRAL API
  description: |
    Ga & Dangme online social community API. One backend for Website, Android and Windows.
    Auth: JWT bearer access tokens (15 min) + rotating refresh tokens (30 days).
    Google sign-in uses genuine OpenID Connect ID tokens verified against Google JWKS.
    Authentication emails are sent from GADAVIRAL <admin@gadaviral.com>.
    Full endpoint reference: docs/API.md
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /auth/register:
    post: { tags: [Auth], summary: 'Register (email+password) → verification link + OTP email', responses: { '202': { description: Sent }, '400': { description: Invalid } } }
  /auth/verify-email:
    get: { tags: [Auth], summary: 'Verify via single-use link (?token=)', responses: { '200': { description: OK }, '410': { description: Expired/used } } }
...
  /auth/login:
    post: { tags: [Auth], summary: 'Login → access+refresh tokens', responses: { '200': { description: OK }, '401': { description: Bad credentials } } }
...
  /auth/change-password:
    post: { tags: [Auth], summary: Change password (revokes sessions), security: [{ bearerAuth: [] }], responses: { '200': { description: OK } } }
  /auth/change-email:
    post: { tags: [Auth], summary: Request email change, security: [{ bearerAuth: [] }], responses: { '200': { description: OK } } }
...
  /auth/me:
    get: { tags: [Auth], summary: Current user + profile, security: [{ bearerAuth: [] }], responses: { '200': { description: OK } } }
...
  '/users/{username}':
    get: { tags: [Users], summary: Public profile, parameters: [{ name: username, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /events:
    get: { tags: [Community], summary: Upcoming events, responses: { '200': { description: OK } } }
    post: { tags: [Community], summary: Create event, security: [{ bearerAuth: [] }], responses: { '201': { description: OK } } }
  '/events/{slug}/rsvp':
    post: { tags: [Community], summary: RSVP, security: [{ bearerAuth: [] }], parameters: [{ name: slug, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /admin/users:
    get: { tags: [Admin], summary: 'User management (admin+)', security: [{ bearerAuth: [] }], responses: { '200': { description: OK } } }
...
components:
...
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }